adjusted:	float becomes double to help reduce artefacts
	modified:   hdr/dataRT.h
	modified:   hdr/miniRT.h
	modified:   scene.rt
	modified:   src/ft_deadzn.c
	modified:   src/ft_det3x3.c
	modified:   src/ft_distce.c
	modified:   src/ft_invmat.c
	modified:   src/ft_keyhook.c
	modified:   src/ft_multlv.c
	modified:   src/ft_newcyl.c
	modified:   src/ft_newlig.c
	modified:   src/ft_newsph.c
	modified:   src/ft_norm.c
	modified:   src/ft_nrmlze.c
	modified:   src/ft_pars_coord_rgb.c
	modified:   src/ft_polyd2.c
	modified:   src/ft_raytra.c
	modified:   src/ft_render.c
	modified:   src/ft_scalar.c
	modified:   src/ft_setamb.c
	modified:   src/ft_setcam.c
	modified:   src/ft_setvec.c
	modified:   src/ft_shades.c
	modified:   src/ft_shdcyl.c
	modified:   src/ft_shdpla.c
	modified:   src/ft_shdsph.c
	modified:   src/ft_syscyl.c
	modified:   src/ft_syspla.c
	modified:   src/ft_syssph.c

+++ src/ft_distce.c
@@ -6,12 +6,12 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/07 18:24:03 by alfux             #+#    #+#             */
-/*   Updated: 2022/12/22 09:33:57 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/28 23:59:48 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <miniRT.h>
 
-float	ft_distce(t_vec a, t_vec b)
+double	ft_distce(t_vec a, t_vec b)
 {
 	return (ft_norm(ft_dif_uv(a, b)));
 }
